requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2-binary
python-jose
python-multipart
alembic
